Add Semantic-ITC dataset support

+++ dataloader/dataset_semantickitti.py
@@ -93,7 +93,7 @@ class voxel_dataset(data.Dataset):
         intervals = crop_range / (cur_grid_size - 1)
         if (intervals == 0).any(): print("Zero interval!")
 
-        grid_ind = (np.floor((np.clip(xyz, min_bound, max_bound) - min_bound) / intervals)).astype(np.int)
+        grid_ind = (np.floor((np.clip(xyz, min_bound, max_bound) - min_bound) / intervals)).astype(int)
 
         # process voxel position
         voxel_position = np.zeros(self.grid_size, dtype=np.float32)
@@ -118,6 +118,7 @@ class voxel_dataset(data.Dataset):
             return_fea = return_xyz
         elif len(data) == 3:
             return_fea = np.concatenate((return_xyz, sig[..., np.newaxis]), axis=1)
+        return_fea = return_fea.astype(np.float32)
 
         if self.return_test:
             data_tuple += (grid_ind, labels, return_fea, index)
@@ -244,7 +245,7 @@ class cylinder_dataset(data.Dataset):
         intervals = crop_range / (cur_grid_size - 1)
 
         if (intervals == 0).any(): print("Zero interval!")
-        grid_ind = (np.floor((np.clip(xyz_pol, min_bound, max_bound) - min_bound) / intervals)).astype(np.int)
+        grid_ind = (np.floor((np.clip(xyz_pol, min_bound, max_bound) - min_bound) / intervals)).astype(int)
 
         voxel_position = np.zeros(self.grid_size, dtype=np.float32)
         dim_array = np.ones(len(self.grid_size) + 1, int)
@@ -267,6 +268,7 @@ class cylinder_dataset(data.Dataset):
             return_fea = return_xyz
         elif len(data) == 3:
             return_fea = np.concatenate((return_xyz, sig[..., np.newaxis]), axis=1)
+        return_fea = return_fea.astype(np.float32)
 
         if self.return_test:
             data_tuple += (grid_ind, labels, return_fea, index)
@@ -344,7 +346,7 @@ class polar_dataset(data.Dataset):
         intervals = crop_range / (cur_grid_size - 1)
 
         if (intervals == 0).any(): print("Zero interval!")
-        grid_ind = (np.floor((np.clip(xyz_pol, min_bound, max_bound) - min_bound) / intervals)).astype(np.int)
+        grid_ind = (np.floor((np.clip(xyz_pol, min_bound, max_bound) - min_bound) / intervals)).astype(int)
 
         voxel_position = np.zeros(self.grid_size, dtype=np.float32)
         dim_array = np.ones(len(self.grid_size) + 1, int)
@@ -367,6 +369,7 @@ class polar_dataset(data.Dataset):
             return_fea = return_xyz
         elif len(data) == 3:
             return_fea = np.concatenate((return_xyz, sig[..., np.newaxis]), axis=1)
+        return_fea = return_fea.astype(np.float32)
 
         if self.return_test:
             data_tuple += (grid_ind, labels, return_fea, index)
@@ -395,7 +398,7 @@ def nb_process_label(processed_label, sorted_label_voxel_pair):
 
 def collate_fn_BEV(data):
     data2stack = np.stack([d[0] for d in data]).astype(np.float32)
-    label2stack = np.stack([d[1] for d in data]).astype(np.int)
+    label2stack = np.stack([d[1] for d in data]).astype(int)
     grid_ind_stack = [d[2] for d in data]
     point_label = [d[3] for d in data]
     xyz = [d[4] for d in data]
@@ -404,7 +407,7 @@ def collate_fn_BEV(data):
 
 def collate_fn_BEV_test(data):
     data2stack = np.stack([d[0] for d in data]).astype(np.float32)
-    label2stack = np.stack([d[1] for d in data]).astype(np.int)
+    label2stack = np.stack([d[1] for d in data]).astype(int)
     grid_ind_stack = [d[2] for d in data]
     point_label = [d[3] for d in data]
     xyz = [d[4] for d in data]
